Update staging.txt
I'm not sure if I'm doing this correctly, but the indicated repo is a full introductory statistics textbook written in bookdown, which uses jamovi for all practicals and examples. The book is in the final stages of copy-editing and will be published with CRC Press and should be available in print in October.